backend/requirements.txt
fastapi
uvicorn
torch
torchvision
opencv-python-headless
pillow
numpy
groq
grad-cam
scikit-learn
python-multipart
python-dotenv
